refactor(exam.payload)
