+++ src/main/kotlin/com/cn/langujet/actor/exam/payload/ExamDto.kt
@@ -1,4 +1,4 @@
-package com.cn.langujet.actor.exam.payload.dto
+package com.cn.langujet.actor.exam.payload
 
 import com.cn.langujet.domain.exam.model.nested.Difficulty
 import com.cn.langujet.domain.exam.model.Exam
