+++ src/main/kotlin/com/cn/langujet/actor/exam/payload/ResultDto.kt
@@ -1,4 +1,4 @@
-package com.cn.langujet.actor.exam.payload.dto
+package com.cn.langujet.actor.exam.payload
 
 import com.cn.langujet.domain.exam.model.Result
 import com.cn.langujet.domain.exam.model.SectionResult
